FileSystemPathTest : Fixed testModificationTimes().
I'm not sure I've ever packed so many bugs into one line before. First off, I was subtracting "now" from a time known to have occurred previously, which would naturally give us a negative number. You might think that therefore the test would always pass, but it failed in Canada. This was because the FileSystemPath class returns UTC time, but python's `datetime.now()` returns local time. So in timezones with a positive offset from UTC we were comparing a large negative number of seconds, and in timezones with a negative offset in UTC we were comparing a large positive number of seconds, and getting a failure. Finally, the threshold for the comparison was smaller than a second, but the filesystem on OS X only has second resolution, so the threshold needed raising.

Also documented the available FileSystemPath properties, including the fact that UTC time is used, to help avoid similar mistakes in the future.