PlugLayout : Added summary tooltips to tabs.
These are identical in nature to the summaries shown in the headers of collapsible sections, and are generated from the same metadata.

Fixes #332. This ticket asked for description metadata to be displayed, but since tabs are layout items rather than plugs, only the section summary actually makes sense. I think the ticket dates back to a time when we didn't have section summaries.