Metadata : Fixed inconsistent handling of NULL values.
There were inconsistencies between the type-based registrations and the instance-based registrations. Type-based registrations of a NULL value could already be used to override a higher level registration, but a registration of NULL for an instance was ignored entirely as a lazy way of implementing the undo for instance values. This then led to NULL values being serialised pointlessly after undoing a registration.

The fix is to treat the explicit registration of a NULL instance value as meaningful in the same way as for a type-based value, and to use boost::optional internally to distinguish between a set-as-NULL value and a not-set-at-all value.