PlugValueWidget : Added support for "plugValueWidget:type" metadata.
This can be used to control the type of widget returned by `PlugValueWidget.create()`. Previously we had a similar mechanism implemented in PlugLayout, but that is of limited use, because other UI elements sometimes use `PlugValueWidget.create()` directly, and we want them to get customised results too. This new mechanism also matches the one added recently for `Nodule::create()`.

Backwards compatibility with the old PlugLayout style metadata is provided for now, but accompanied by a warning.