UIEditor : Made labels fixed width.
It was a bit jarring jumping between having a section selected and a plug selected, because the label column in the panel on the right changed width. This also allows us to switch from using a GridContainer to using ListContainers, so the structure of the UI is much more evident from the indenting in the code, and we don't need to faff around with parenting parameters full of indices that need changing whenever you insert something.