Added a Loop node.
This takes an input and loops it N times through an external graph before outputting it again. This provides the user with the ability to do things with the graph which were previously only achievable with code.

The implementation follows the pattern established by classes like Switch - a generic template class makes it possible to mix-in the loop behaviour on top of an existing processor base class. In addition, instantiation as LoopComputeNode allows any data type to be processed by adding appropriate in and out plugs after construction - this form is demonstrated by the unit tests.