Plug : Added AcceptsDependencyCycles flag.
This will be used to allow dastardly nodes to do dastardly things. In actual fact, we already have nodes which deliberately cause cycles during compute - for instance we call `outPlug()->bound( someChild )` from within implementations of `SceneNode::computeBound()` in order to recursively propagate bounding boxes up from the lower levels of a scene. Those existing examples were all of direct cycles on a single plug though, and we would now like to do the same but involving relationships between several plugs, which requires the cyclic dependency to be declared via `DependencyNode::affects()`. This flag allows that to be done.