ImageTransform : Fixed copy + paste.
The names of the internal implementation nodes were inexplicably being based on the name of the node at the point of construction. This meant that if the node was then renamed and cut + pasted, the serialised node was referring to the implementation details via the original name, and not the new one. Having a fixed name for the internals is the only sensible thing in any case - I'm not sure why we went out of our way to create this bug.

Also simplified connection logic.