Parallelized CopyTiles by channelName as well.
This only shows very mild speed improvements on the large images (scenario A,C,D), as those are threading fairly well already, but the benefit can been seen when using very small images. Prior to this, scenario B only used a single thread (because it has a single tile for the output image), but now it uses 4 threads during the full compute.

A) one reformat from 8k -> 4k
B) a downsample chain of reformats from 8k -> 1x1
C) the chain from B followed by an upsample chain from 1x1 -> 8k
D) a double loop setup with the same effect as C

For each of the above scenarios I'm calling node.out.image() on the last node in the setup, twice. This is because the first call is bound by the ImageReader (and actually OpenImageIO), so the second call gives us a better idea of the hash() performance.

Here are the timings:

A)
before	5.195990	0.163803
after	5.153800	0.142413

B)
before	27.682336	0.000241
after	8.034444	0.000320

C)
before	57.058365	1.889977
after	56.891124	1.795154

D)
before	50.549267	4.561736
after	49.855733	4.212652