Gaffer : Added Wedge node.
This causes upstream ExecutableNodes to be executed multiple times with a range of values specified by a context variable. By referencing this variable in upstream expressions, it's possible to easily produce renders with a wedged parameter range. The variable can be specified as an integer range, a float range, a colour ramp or explicit lists of ints, floats or strings. Wedge nodes may be connected in series to perform multidimensional wedges with more than one parameter.