AppleseedRender : Simplified output of the world.
There's not much point using a ScriptProcedural to load the current script separately if we're going to just execute it immediately anyway - we can just use a SceneProcedural with the current script itself. This also has the side benefit that wedge renders can now be performed, since the old method failed to transfer context values other than the frame number into the render process, but this method uses the current context directly.