FileMenu : Worked around PySide crashes when opening recent files.
This appears to be the same issue the original workaround was for, but triggering itself only when an error occurred during loading. I think this is because an idle event triggers while we display the "Oy Vey" dialogue, so we must wait until after that dialogue has closed to schedule the removal of the script.