Reference : Removed "fileName" plug.
As per #801, this allows the user to create a plug called "fileName" on a Box, and have it referenced correctly onto a Reference node. Also added Reference::referenceLoadedSignal() so that the current reference can be tracked by observers, and add test cases for undo/redo of reference loading.

Backwards incompatibilities :

- Removed Reference::fileNamePlug() method.
- Removed Reference "fileName" plug.
- Old scripts containing Reference nodes will require `continueOnError = True` to load.