Expression : Reused existing plugs/connections where possible.
When an expression was set, we were always rebuilding our connections to the graph, even if the connections we already had were exactly what we needed. This means we no longer do redundant work when loading serialised expressions. This fixes a problem found using expressions with the ImageWriter node in production.