ImageWriter : Removed unwarranted "Invalid filename" exceptions.
These occurred in the following situations :

- When undoing the filename back to ""
- When using substitutions so that the filename is valid only in specific contexts
- When using an expression so that the filename is valid only in specific contexts

The ImageWriter was trying to be smart and make the mode plug read-only in the event that a format doesn't support tiles. But the concept of a "current filename" really doesn't exist in a multithreaded app with different contexts per thread, so this idea really doesn't make sense. There is logic in execute() already to fall back to scanline mode when necessary, and the documentation mentions this behaviour.