Wedge : Replaced floatStep plug with floatSteps plug.
Rather than define an increment to be applied repeatedly until the max value is reached, we now defines a number of steps to divide the min-max range into. This is more robust because the max value is guaranteed to be included in the range, whereas even with a seemingly simple range defined by 0-1 in increments of 0.1, the max can be omitted due to floating point precision issues.

The int range continues to be defined in terms of an increment, because defining in terms of a number of steps would allow nonsensical non-integer increments to be specified.