ValuePlug : Emitted plugSetSignal() when children added/removed.
I had hoped we might not need this, but it turns out to be necessary for the CompoundDataPlug as well as the SplinePlug, so having it on the base class makes sense. Since we avoid connecting our handler in the leaf plug case, this leads to no more connections than we had previously.