PlugLayout : Activators drive enabled-ness rather than read-only-ness.
Whether a plug (or UI) is read-only or not is orthogonal to whether or not the plug is relevant given the current values of other plugs. Since activators were intended to model the latter concept, it makes more sense to use the enabled state of the widget to implement their effect.