SceneNode : Improved bound propagation performance.
The `ScenePlug::bound()` and `ScenePlug::transform()` convenience methods both create a temporary context internally. We can therefore reduce the number of context creations by avoiding the convenience methods and making our own context once outside of the loop.

This shaves 7% off the time to propagate bounds when a Transform node is used with a particularly inefficient filter.