GafferImage : Added a Shuffle node.
This shuffles the data from one channel into another, for an arbitrary number of channels. It ensures that data is never copied and that the Shuffle and its input node will share cache entries. Synthetic white and black channels may be used as the source for a shuffle - this is particularly useful for creating a solid alpha channel when a one doesn't exist yet.

Fixes #1380.