ExecutableNode : Support promotion of requirements ArrayPlug.
This required the replacement of the generic Plugs we were using to represent requirements with a RequirementPlug, so we could use typing for our `acceptsInput()` implementation. This is much preferable to our previous form anyway, since it avoids the nasty "if it's a box or a dot then let's be lenient and hope for the best" approach. We should adopt this approach in similar situations (filter and shader assignment plugs).