Promoted plug names are now relative to the node they came from.
This somewhat simplifies the verbose names that we were getting before, and it allows us to compute the name prior to adding the plug to the Box, so we can make use of it in Box::create() without multiple re-names.

It also avoids confusion which can occur when promoting a plug and then renaming the underlying node it was promoted from.

It does however leave the promoted plugs somewhat ambiguous if several of the same name are promoted, but that seems something the UI at the Box level should tackle (perhaps with a tool tip listing embedded connections), rather than via the plug name itself.