ImageProcessor : Added option of having an ArrayPlug of image inputs.
The original idea for this was that all ImageProcessors would have an ArrayPlug containing the inputs, even though many of them would fix `minSize==maxSize==1`. This was rejected on the grounds that it made for a very awkward Python API, where the single input would have to be accessed with `node["in"][0]` rather than the current concise `node["in"]`.

Another idea for providing similar functionality was to have a separate MultiImageProcessor which would always have an ArrayPlug. This was rejected on the grounds that it would lead to needing MultiChannelDataProcessor, MultiColorProcessor etc etc etc.

This new approach keeps exactly the same API (both C++ and Python) for single-input nodes, while adding the possibility of also having multiple-input nodes using the same base class with an ArrayPlug. The existing `inPlug()` C++ method is implemented to remain useful even for the ArrayPlug case, and the new `inPlugs()` method provides access to the ArrayPlug itself if it exists.