Merge : Added compatibility config for files from previous versions.
This is achieved by redirecting the lookups for `node["inN"]` to `node["in"][N]`, and injecting into `node["in"]` so that it redirects to `node["in"][0]` when it appears that the caller is assuming it is still an ImagePlug rather than an ArrayPlug.

All tests pass now, except `GafferImageTest.MergeTest.testAffects()`, which only fails because a new parent plug has been introduced and that is also listed in the dirtied plugs.