Travis : Workaround GitHub API query limit.
We've been getting our Travis setup failing to launch fairly frequently recently, and it turns out it was the API queries we were making to figure out the latest gafferDependencies release. Unless you authenticate with some credentials, there's a limit of 60 queries per IP address per hour, and it seemed that often that limit is reached by whatever occurred on the Travis VM before our tests launched. One option would be to provide some credentials, but there's not really a good way of keeping those private, so we instead hardcode the path to the dependencies package directly.