Remove Cortex from the node menu.
Many of the Ops have names which clash with built in nodes, and it's infuriating to accidentally drop down a useless curves merge op when you wanted a GafferImage::Merge node. I see Cortex increasingly just taking a "behind the scenes" role in Gaffer, with it being much more natural for a TD to use a Box or Python to design a custom node than to write an Op.