Switch : Reimplemented to use ArrayPlugs.
All GafferTest.SwitchTest tests pass with the minor modifications made. Note that unlike the previous InputGenerator->ArrayPlug changes, this one does not support backwards compatibility with files from previous versions, because the ArrayPlug is not added in a constructor, and is instead added as a dynamic plug. As far as I can tell, we have not yet used such a Switch in anger (it is not exposed via the UI).

Backwards incompatibility :

- SwitchComputeNode and SwitchDependencyNode now require that the "in" plug is an ArrayPlug.