Config : Reduced impact of InputGenerator compatibility.
We no longer inject a new __getitem__ method onto the Node base class, instead injecting it only onto the subset of classes which require it. We continue to require injecting methods onto all ArrayPlugs, because it's not possible to inject __getitem__ on a per-instance basis (only on a per-class basis).