Sampler binding : Fixed binding of sample( int, int ).
Boost resolves in overloads in reverse order of registration, so we have to bind int last otherwise calls with ints are swallowed by the float version. Also added some more tests for the sample methods.

Really though, the sample methods should have been taking V2i and V2f all along - hopefully we'll deal with that in the not too distant future.