GafferImage : Add Resample node.
This is an internal utility node that resamples the data from one display window into another, resizing, offsetting and filtering as necessary. It will be used in the internal implementations of other nodes. It uses OIIO::Filter2D rather than GafferImage::Filter, since the former has a more complete set of filters, plus the ability to have non-separable filters and to distinguish them from separable ones. If this works out, we'll be aiming to remove GafferImage::Filter entirely.