Resample : Add filterWidth plug.
The intention is that this will allow the Resample node to be used in the internal implementation of a Blur node. Currently the brute-force nature of the filtering algorithm means that a 100x100 gaussian blur of an HD image takes a whopping 254 seconds. This will be addressed in subsequent commits.