ScriptNode : Save version metadata in `serialiseToFile()`.
This can be used for informational purposes and also to assist in providing backwards compatibility with old files. This mechanism has been in use for some time in the Box/Reference export code - this just generalises it for all serialised files. At the time we introduced it for Boxes I was unsure if it should apply elsewhere, but we now have a use case for it on ScriptNodes (Hugh's #1444). Since serialising to file implies later reloading that file, it seemed worth providing this information everywhere a file is saved, rather than just in `ScriptNode::save()`.

When importing a file into an existing ScriptNode, the values of the version metadata are potentially a little confusing, since the import overrides the existing value. The metadata is therefore best viewed as "the version of gaffer used to write the last file that was loaded or imported". Since our primary use case is during loading, hopefully this is acceptable for now.

Fixes #1436.