UIEditor : Never edit user plug for Box nodes.
We only want the user to edit plugs directly on the box, so that the user plug is available to downstream users of References (and one day gizmo type things). But we were trying to provide help for old boxes by editing the user plugs instead if they'd been used previously. This isn't working out, so we're removing that fallback, and will provide a separate means of upgrading old boxes internally at IE.