Sampler : Fix broken clamp bounding mode.
It would cause crashes in the event that the requested sample window and available data window didn't intersect. Furthermore, it would tolerate requests outside the requested sample window and silently clamp them back inside it. This masks bugs, since providing the appropriate sample window is essential to get the right result from `Sampler::hash()`.

We fix this by requiring that the client code provides the right sample window on pain of undefined behaviour (debug assertions assist in validating this code). We can then clamp consistently only into the data window, and arrange that we have cached the pixels within this clamp target even if they are outside of the originally requested sample area.