Sampler : Fix hash() to account for pixels outside data window.
The hash is only useful if it represents the values for all pixels that might be queried from the sample area, so we must include the sample and data windows to account for black (or clamped) pixels within that area.

Also added a convenience overload of hash() for the case where you are querying the hash rather than appending to an existing hash.