Expression : Hide Engine from public API.
As the Engine API becomes more complex, we want to make it clear that it is entirely an implementation detail as far as clients of the Expression class go. The Expression node will provide a simpler API for just the tasks a client needs. The Engine remains exposed only so much as is necessary to allow subclassing to add languages.