Plug : Fixed bug in child connection tracking behaviour.
When all the children of a plug are connected to all the children of another plug, we automatically make the equivalent connection at the parent level, so that the parent connection state accurately reflects that of the children. However, we were rather overzealous in connecting the parent, doing it in several situations where the child connections were not actually equivalent to those that would have been made by connecting the parent directly.

If a parent has an incoming connection, it now guarantees that the children are in the same state they would be had `parent->setInput( otherParent )` been called directly.

This fixes ExpressionTest.testYDrivingX().