Compatibility : Fixed loading of UnionFilters from Gaffer 0.15.0.0.
The UnionFilter used to make its own ArrayPlug input, and named the first element "in". When we introduced the FilterProcessor class, it named the first element "in0". So when we rederived UnionFilter from FilterProcessor we broke backwards compatibility with old files. We restore it by detecting failed element lookups and redirecting them.

This also fixes a problem where we were trying to reraise an exception in the compatibility config but where doing so by referencing a nonexistent variable `e`. Since all we did in the handler was reraise, we can actually just remove the handler.

Fixes #1474.