Compatibility : Fixed loading of FilterSwitches from Gaffer 0.15.0.0.
There were two problems here. Firstly, the setInput() call for the "out" plug appears in the serialisation before the setInput() calls for the "in" plugs. We therefore must transfer the old InputGenerator plugs into the ArrayPlug rather than throw them away, so that the inputs exist in time. Secondly, if the index is 0, then we have a serialised `node["out"].setInput( node["in"] )` call, but "in" is now an ArrayPlug. So we connect the first child of the array instead.