Context : Introduce "framesPerSecond" variable and time accessors.
Till now we've had no official concept of a time other than the current frame. By introducing a "framesPerSecond" variable into every Context, we can map the frame to and from a time in seconds. Note that we deliberately don't also introduce a third "time" variable as that would require synchronization between the dependent variables when they are set.