OpenGLShaderTest : Add expectedFailure decorator on OSX.
This is the failure reported in #1480. There is a workaround we could use, which is described in the comment, but since the real problem is elsewhere and this has yet to cause a problem in the real world, I've gone for just documenting the issue in the hope that one day it'll be fixed in the compiler.