GraphGadget : Support dynamic addition/removal of Nodules.
This is needed now that the user can use the UIEditor to add and remove nodule metadata from plugs on Boxes. When nodules are added/removed the GraphGadget needs to make sure that the corresponding connections are updated appropriately.

To help in the implementation, the `addConnectionGadgets()` and `removeConnectionGadgets()` methods now operate on the connections for the passed plug itself, rather than just on the children.