Added plugsWithMetadata and nodesWithMetadata to Gaffer::Metadata
I'm doing some asset management stuff where I flag assetized StringPlugs (eg plugs pointing to caches on our asset management system) with a special metadata key. The asset management system needs to scan for all plugs with this data in the scene, and it seems this is the most efficient way of doing that. I haven't tried a brute force search in python, but a brute force search in c++ takes 0.1 seconds in a reasonably sized scene I tried. That latency's kind of ok, but I wouldn't be surprised if I saw scenes ten times as big and a 1 second latency could get pretty annoying. On the other hand, if I use plugsWithMetadata with the special instanceOnly case, the search takes 1.e-5 seconds for my test scene, which is a lot better.