ImageStats : Update for default format API changes.
Also use a post create method in the UI rather than have the node modify its own plug values. It is not expected that nodes in Gaffer try to be "helpful" in this manner, because they have no way of knowing what context to be helpful in, and it makes the API less predictable if changing an input has unexpected side effects.