ImageTest : Add ImageTestCase with assertImagesEqual() method.
This will allow us to standardise on a single way of comparing images, and move away from IECore.ImageDiffOp (since we want to remove most of the image functionality from cortex in preference to OIIO). It also has the nice feature of actually telling you what the maximum difference is rather than just a true/false result like ImageDiffOp.