Crop : Fix affects().
- Input dataWindow and format weren't affecting the equivalent outputs.
- The affectsDisplayWindow and affectsDataWindow plugs were each affecting both.

I've gone for a very simple implementation where each output that could be dirtied has its own conditional block, even if that means some duplicate tests. It just seems too easy to get the logic tangled for anything complex. As noted in a previous commit, the logical extension to this may well be a set of `virtual affects*()` methods to match the existing `hash*()` and `compute*()` methods.