Reference : Support metadata edits.
Previously we never serialised _any_ metadata from Reference nodes, regardless of whether the metadata originated from within the referenced file or was added/edited after loading. We now support both the addition of new metadata and editing of metadata from the reference.

This implementation uses the persistence flag on each metadata item to track whether it is considered to have come from the referenced file or not. Strictly speaking, this is slightly abusing the semantics of that flag, but this does allow us to support edits with minimal changes, and is similar to the way we treat the Dynamic plug flag already. We still need a full system for tracking reference edits to plug values - when we implement such a thing we should probably track metadata edits with the same mechanism.

Fixes #1536.