Workaround for a shortcut key problem
It turns out that shortcuts like Del etc are completely broken when people dock the Caribou gaffer window in a maya panel. This is because the shortcut keys are scoped on the entire window, and the gaffer ones start overriding the maya ones. I found a way of scoping the shortcuts so it takes the current focus into account and this doesn't happen (see comments), which I've used to put together a slightly sketchy looking workaround for the problem. There's probably a better way of doing it, so this is more of a bug report than a pull request really.