Merge : Fix data window when first input unconnected.
Just because we only have one input, it doesn't mean we're disabled, because the input we have might not be to `inPlugs()[0]`.

Also tidied up the implementation for hashDataWindow() and computeDataWindow(), to not check the same plug twice, and to avoid querying the existence of input connections (which ComputeNodes should be unaware of).