ImageNode : Deunoptimise hash() method.
Calling `ancestor<ImagePlug>()` would traverse all the way to the root node in the case that no such ancestor is found, calling `isInstanceOf()` for every node on the way. This is a quick enough operation for C++ nodes, but for nodes implemented in python the subsequent locking before calling the python override is a total performance killer. And it's actually fairly common to have Box subclasses implemented in python which then contain image nodes. Since we know that ImagePlug only has children one level deep, we can simply use `parent<ImagePlug>()` to avoid the problem.