ExecutableNode : Rename requirements to tasks.
This is motivated by the consensus that the "requirement" plug in particular is confusingly named, and that preTasks and postTasks are more intuitive than requirements and someOtherNameWereNotSureAbout.

A compatibility config file maintains backwards compatibility. DispatcherTest needed tweaking in one spot where it was making assumptions about the names that would be given to promoted plugs.