Dispatcher : Refactor batching into dedicated class.
The batching utility methods all needed to pass shared state explicitly between themselves, and this was getting rather awkward, particularly considering that upcoming features would require more state to be passed. By making a utility class we are able to store the state as member data, and closely associate it with the batching methods.

Nothing has changed about the dispatching logic here, it's literally just shuffling the original code into an easier-to-extend form.