Dispatcher : Refactor Batcher::acquireBatch() for clarity.
Previously the logic for adding frames was in two places, depending on whether or not we were using an existing batch (where we explicitly added a frame) or making a new one (where it was implicitly added by the TaskBatch constructor). Having a single spot where we update the batch will be important as we add more state to batches in the future.

Also, make local variables const where possible.