LocalDispatcher : Removed "executeInForeground" ExecutableNode plug.
This is equivalent to the new "immediate" plug provided by the Dispatcher base class. Note that no backwards compatibility is being provided here at all. We could perhaps provide some via a config file redirecting to the immediate plug, but if possible I'd prefer a clean break.

We should also discuss the possibility of removing the foreground mode from the LocalDispatcher entirely, and instead having a global immediate mode provided by the Dispatcher base class. The way the LocalDispatcher has entirely different code paths for foreground and background dispatch is ripe for errors anyway, so it might well make sense to simplify things this way.