PythonCommand : Support sequence execution.
This is requested via a new bool plug called "sequence". When enabled, a list of frames is available to the command, and the command may access the value of a variable at any valid frame by updating the context before accessing the variable.

To implement the "per-frame variable values" the variables were moved from `variableName` to `variables["variableName"]` within the command syntax. It would have been possible to maintain the old form, but it would have had these pitfalls :

- We would have suffered from the inconsistent python behaviour described here :
	- https://mail.python.org/pipermail/python-dev/2002-October/029742.html
	- ScriptNodeTest.testClassScope()
	- http://bugs.python.org/issue991196
- It would be possible to enter a variable name via the UI that is not a valid python name