ValuePlug : Fix hang triggered by plug deletion during compute.
This problem was distilled down from a test hang diagnosed and
described in #1576 :

> There's a problem in the core caching code, whereby we hit a
> lock if a plug is deleted in a thread that is currently performing
> a graph computation. Gaffer's threading model has the constraint
> that a particular thread is allowed to either be modifying a node
> graph or computing with one, but not doing both at the same time.
> So you can construct a graph and then hammer it with computations
> in multiple threads, or compute with one graph in some threads while
> constructing or destructing a different graph in another thread.
> This is a pretty reasonable constraint, because it doesn't make any
> sense to mutate a graph while computations are being performed using
> it. And it really doesn't take any effort to meet this constraint in
> C++, because to break it you'd have to really go out of your way to
> do something odd.
>
> It turns out Python is a different matter though, because of the
> garbage collector, and that's what's causing the hang. The hanging
> test includes a python expression, so the computation is entering
> python from C++ to evaluate it. Although we only want to be doing
> a compute, python decides that now is a great time to perform a
> garbage collection cycle, and during garbage collection it's
> deleting plugs from an old graph that was constructed by an earlier
> test. Which gets us into the situation where we're inadvertently
> modifying a graph from a thread that we're currently computing in.

When this happened, we were locking the thread's hashCacheMutex at
the start of computation (in the Computation constructor), but then
trying to lock it again in clearHashCaches(), before the computation
was complete.

A couple of potential solutions were rejected :

1. Disable garbage collection when entering python during a compute.
   Rejected on the grounds of fragility, and difficulty of implementation
   given the number of places we enter python from C++.
1. Make hashCacheMutex recursive, so it accepts the second lock.
   Rejected because recursive mutexes are slower than non-recursive
   ones, and it turns out we don't need a mutex at all.

The final solution is to simply defer the clearing of the cache until
the start of the next computation in each thread. This is OK because it
is illegal to modify a graph while computations are being performed
with it, and we know we will have cleared the cache in time for the
next computation.

This should leave us with an even simpler threading constraint, where
we no longer say "thou shalt not modify _any_ graph while computation
is being performed", but instead say "thou shalt not modify a graph
while _it_ is being used for computation".