PythonCommand : Add frame to hash automatically in sequence mode.
It's confusing to turn on sequence mode and then have to add some other frame dependency by hand to actually be given the full list of frames. While this does mean that the opportunity to have identical frames collapsed down into a single frame by the dispatch engine is gone, it seems unlikely that the typical user of the PythonCommand cares about this more than basic usability. If we later find we do have a use case for the fine distinction, we can add a "trust me i know what i'm doing" plug.