MenuBar : Fix handling of shortcuts with nested menubars.
We want to allow a single window to have multiple menubars with identical shortcuts, without them conflicting. This is made tricky by the fact that Qt likes to scope the shortcuts at the window level, where they are bound to conflict. We work around this by doing manual shortcut processing on the parent of the menubar. This has the effect of scoping menubar shortcuts such that they are accessible within the menubar's parent (and all it's child widgets) but nowhere else.