GafferImage : Fix reference counting.
When calling `TypedObjectPlug::getValue()`, the caller is responsible for holding the return value with a ConstPtr for as long as the contents are needed. Otherwise the value can die while it is still being accessed. I'd like to consider ways we could check for this sort of gotcha automatically on Travis at some point...