StandardNodeToolbar : Support toolbars on all edges.
An argument to the constructor chooses the edge to create the toolbar for, and "toolbarLayout:section" metadata variables with values of "Top", "Bottom", "Left" and "Right" choose which toolbar a plug belongs in. The layout is then performed using a PlugLayout rooted at that section.

Also added static utility methods to create a toolbar for each edge - these are useful when registering toolbar types via metadata. For instance, "nodeToolbar:top:type", "GafferUI.StandardNodeToolbar.top".