SceneProcedural : Optionally use IECore::Renderer::Procedural::noBound.
For some graphs, computing the procedural bounds is a significant chunk of the runtime. Furthermore, in many rendering scenarios the bounds are not useful because raytraced global illumination means that all procedurals will be opened pretty quickly. In fact, 3delight pretty much requires that all procedurals are evaluated before rendering starts in anything but REYES mode anyway. In this scenario, our motivation for using procedurals is really just to allow threaded scene generation and not for the lazy-evaluation-on-demand of days gone by. Hence, it makes sense to have the option to avoid evaluation of bounds entirely.